backend/requirements.txt
# Core Framework
pydantic
fastapi
uvicorn[standard]

# Database
pymongo
motor

# AI/ML
google-genai
openai
numpy

# Utilities
python-dotenv
httpx

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart